feat: add ability to add items to cart.
Cart Context has been modified to include the addItem function. This
function takes in an item and adds it to the cartItems array. If the
item already exists in the cart, it will increase the quantity of that
item by 1. If the item does not exist in the cart, it will add the item
to the cartItems array with a quantity of 1.

The Cart Dropdown component has been modified to include the Cart Item
component. This component will display the item name, quantity, and
price. It will also display the total price of the item.