feat: add header component.
Header component has been added to the project. It contains a logo and
a button. The button is a placeholder for creating notes and will be
implemented in the future.

Logo component has been created to be used in the header component.

Button component has been created to be used in the project.